Move the DownloadProgressIterable to the helper module
We are going to re-use this class for other downloads too.
