+++ pontos/helper.py
@@ -17,7 +17,54 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import subprocess
-from typing import Iterable
+from pathlib import Path
+from typing import Iterable, Iterator, Optional
+
+
+class DownloadProgressIterable:
+    def __init__(
+        self, content_iterator: Iterator, destination: Path, length: int
+    ):
+        self._length = None if length is None else int(length)
+        self._content_iterator = content_iterator
+        self._destination = destination
+
+    @property
+    def length(self) -> Optional[int]:
+        """
+        Size in bytes of the to be downloaded file or None if the size is not
+        available
+        """
+        return self._length
+
+    @property
+    def destination(self) -> Path:
+        """
+        Destination path of the to be downloaded file
+        """
+        return self._destination
+
+    def _download(self) -> Iterator[Optional[float]]:
+        dl = 0
+        with self._destination.open("wb") as f:
+            for content in self._content_iterator:
+                dl += len(content)
+                f.write(content)
+                yield dl / self._length if self._length else None
+
+    def __iter__(self) -> Iterator[Optional[float]]:
+        return self._download()
+
+    def run(self):
+        """
+        Just run the download without caring about the progress
+        """
+        try:
+            it = iter(self)
+            while True:
+                next(it)
+        except StopIteration:
+            pass
 
 
 def shell_cmd_runner(args: Iterable[str]) -> subprocess.CompletedProcess:
